Removed i64 literal suffix.

+++ onnxruntime/test/providers/cpu/ml/array_feature_extractor_test.cc
@@ -47,7 +47,7 @@ TEST(MLOpTest, ArrayFeatureExtractorHigherDimensionalXTest) {
 
   const std::vector<int64_t> x_dims{2, 3, 4, 5};
   const int64_t x_size = std::accumulate(
-      x_dims.begin(), x_dims.end(), 1i64, std::multiplies<>{});
+      x_dims.begin(), x_dims.end(), static_cast<int64_t>(1), std::multiplies<>{});
   const std::vector<int32_t> X = [x_size]() {
     std::vector<int32_t> v(x_size);
     std::iota(v.begin(), v.end(), 0);
